Run GHA workflow on tags too
This will catch if we forget to regenerate the new version doc when we
push a flag (which we can thank yank and properly release).

There was an unused submodule pull in the checkout setup for this step
in the workflow, so that's just been removed (I think bad copy paste).

Also update the step to print any diff to help diagnose issues